Move shared types, update email API and UI
Consolidate shared types into app/NX/types.d.ts (add T_Severity, I_MakeRes, T_Email) and remove the old app/api/types.d.ts. Update imports across API and lib files to reference the new types location (makeRes and email routes). Clean up email endpoints by deleting the duplicate route copy and adjusting route.ts/resend imports. UI tweaks: remove NX admin button from Nav, refactor Virus card markup to use a Box wrapper and move CardHeader into the image branch with a link action. Also restore makeRes baseURL and change tenant config background from #ffd849 to #fff.
