+++ app/NX/types.d.ts
@@ -1,6 +1,31 @@
 import type { T_UbereduxDispatch, T_RootState } from '../NX/Uberedux/store';
 export type { T_UbereduxDispatch, T_RootState }
 
+export type T_Severity = 'success' | 'error' | 'warning' | 'info';
+
+export interface I_MakeRes {
+    severity: T_Severity;
+    message: string;
+    data?: any;
+    other?: any;
+}
+
+export type T_Email = {
+    from: {
+        label: string;
+        email: string;
+    },
+    to: {
+        label: string;
+        email: string;
+    }
+    subject: string;
+    body: string; // Markdown or HTML content?  
+    template?: string;
+
+};
+
+
 export type T_Tenant = 'my-tenant' |
     'new-tenant'
 ;
@@ -180,8 +205,6 @@ export type T_Feedback = {
     title?: string;
     description?: string;
 } | null;
-
-export type T_Severity = 'success' | 'info' | 'warning' | 'error';
 
 export type T_NavItem = {
     title: string;
