+++ app/api/lib/makeRes.ts
@@ -1,4 +1,4 @@
-import { I_MakeRes } from "../types";
+import type { I_MakeRes } from "../../NX/types";
 import { makeTime } from './makeTime';
 import { getBaseurl } from './getBaseurl';
 
@@ -6,7 +6,7 @@ export function makeRes({ severity, message, data, other }: I_MakeRes) {
     const epoch = Date.now();
     const meta = {
         time: makeTime(epoch),
-        // baseURL: getBaseurl(),
+        baseURL: getBaseurl(),
         severity,
         message,
