+++ app/api/notify/email/logs/route.ts
@@ -1,11 +1,7 @@
-import type { T_Email } from '../../../types';
 import { NextResponse } from 'next/server';
 import { makeRes } from '../../../';
 
 export async function GET() {
-
-
-
 
     const res = makeRes({
         severity: 'warning',
